refactor: 베이스 엔티티 개선 (#121)
* refactor: 학기 패키지를 domain에서 support로 이동

* refactor: BaseEntityListener가 semester를 사용하도록 개선

* docs: 주석 추가

* docs: 주석 추가
