+++ src/main/java/kr/allcll/backend/support/entity/BaseEntity.java
@@ -8,8 +8,20 @@ import java.time.LocalDateTime;
 @EntityListeners(value = {BaseEntityListener.class})
 public abstract class BaseEntity {
 
-    String semesterAt;
-    LocalDateTime createdAt;
-    LocalDateTime deletedAt;
-    boolean isDeleted;
+    protected String semesterAt;
+    protected LocalDateTime createdAt;
+    protected LocalDateTime deletedAt;
+    protected boolean isDeleted;
+
+    void setSemesterAtIfAbsent(String semesterAt) {
+        if (this.semesterAt == null) {
+            this.semesterAt = semesterAt;
+        }
+    }
+
+    void setCreatedAtIfAbsent(LocalDateTime createdAt) {
+        if (this.createdAt == null) {
+            this.createdAt = createdAt;
+        }
+    }
 }
