+++ src/main/java/kr/allcll/backend/support/entity/BaseEntityListener.java
@@ -1,29 +1,20 @@
 package kr.allcll.backend.support.entity;
 
 import jakarta.persistence.PrePersist;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import kr.allcll.backend.support.semester.Semester;
 
 public class BaseEntityListener {
 
     @PrePersist
-    public void setSemester(BaseEntity entity) {
-        entity.createdAt = LocalDateTime.now();
-        entity.semesterAt = Semester.now();
+    public void initialize(BaseEntity entity) {
+        entity.setSemesterAtIfAbsent(getCurrentSemester());
+        entity.setCreatedAtIfAbsent(LocalDateTime.now());
     }
 
-    private static class Semester {
-
-        private static final int MIDDLE_OF_YEAR_MONTH = 6;
-
-        public static String now() {
-            LocalDateTime now = LocalDateTime.now();
-            int year = now.getYear();
-            int month = now.getMonthValue();
-            return String.format("%d/%s", year, calculateSemester(month));
-        }
-
-        private static String calculateSemester(int month) {
-            return month <= MIDDLE_OF_YEAR_MONTH ? "1학기" : "2학기";
-        }
+    private String getCurrentSemester() {
+        Semester semester = Semester.getCode(LocalDate.now());
+        return semester.getValue();
     }
 }
